main関数を消した

+++ ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: sshimura <sshimura@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/16 11:43:31 by sshimura          #+#    #+#             */
-/*   Updated: 2024/04/26 14:52:17 by sshimura         ###   ########.fr       */
+/*   Updated: 2024/05/06 14:55:38 by sshimura         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 {
-	unsigned int	i;
-	unsigned int	src_len;
+	size_t	i;
+	size_t	src_len;
 
 	src_len = ft_strlen(src);
 	i = 0;
@@ -30,25 +30,3 @@ size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 	}
 	return (src_len);
 }
-
-//#include <string.h>
-
-//int main(int argc, char** argv)
-//{
-//	if (argc != 3)
-//		return (0);
-//	printf("MINE:\n");
-//	char dst[10];
-//	const char	*src = argv[1];
-//	size_t size = atoi(argv[2]);
-
-//	size_t ft_result = ft_strlcpy(dst, src, size);
-//	printf("The number of characters is %zu\n", ft_result);
-//	printf("ft_dst: %s\n\n", dst);
-
-//	printf("ORIGINAL:\n");
-//	char ft_dst[10];
-//	size_t result = strlcpy(ft_dst, src, size);
-//	printf("The number of characters is %zu\n", result);
-//	printf("dst: %s\n", dst);
-//}
