+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: sshimura <sshimura@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/16 10:35:39 by sshimura          #+#    #+#             */
-/*   Updated: 2024/04/24 11:27:19 by sshimura         ###   ########.fr       */
+/*   Updated: 2024/05/06 14:55:47 by sshimura         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
@@ -23,14 +23,3 @@ size_t	ft_strlen(const char *str)
 	}
 	return (i);
 }
-
-//int	main(int argc, char* argv[])
-//{
-//	if (argc != 2)
-//		return (0);
-//	else
-//	{
-//		int result = ft_strlen(argv[1]);
-//		printf("%d\n", result);
-//	}
-//}
